Consolidate filewriter util.
